requirements.txt
drf-yasg
django-rest-swagger
django-cors-headers